extract common parts
