+++ extensions/terminal-suggest/scripts/pullZshBuiltins.ts
@@ -3,16 +3,11 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { exec } from 'child_process';
-import { promisify } from 'util';
 import * as fs from 'fs/promises';
 import * as path from 'path';
-import { platform } from 'os';
+import { checkWindows, execAsync, copyright } from './terminalScriptHelpers';
 
-if (platform() === 'win32') {
-	console.error('\x1b[31mThis command is not supported on Windows\x1b[0m');
-	process.exit(1);
-}
+checkWindows();
 
 const latestZshVersion = 5.9;
 
@@ -126,14 +121,14 @@ const shortDescriptions: Map<string, string> = new Map([
 	['ztcp', 'Manipulate TCP sockets'],
 ]);
 
-const execAsync = promisify(exec);
-
 interface ICommandDetails {
 	description: string;
 	args: string | undefined;
 	shortDescription?: string;
 }
+
 let zshBuiltinsCommandDescriptionsCache = new Map<string, ICommandDetails>();
+
 async function createCommandDescriptionsCache(): Promise<void> {
 	const cachedCommandDescriptions: Map<string, { shortDescription?: string; description: string; args: string | undefined }> = new Map();
 	let output = '';
@@ -258,11 +253,5 @@ const main = async () => {
 		console.error('Error:', error);
 	}
 };
-
-const copyright = `
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/`;
 
 main();
